feat: Update Instagram image and static subdomains
The commit updates the `instagram.yaml` file to modify the image and static subdomains for Instagram. The `img` subdomain is changed from `instagram.fcdg1-1.fna` to `instagram.flos1-1.fna`, and a new subdomain `static.xxx` is added with the original subdomain `static.xx`. This change aims to improve the performance and reliability of image and static content delivery for Instagram.